Make sure we're closing and deleting our file objects when we're done with them
I saw some odd behavior with redundant `data.tar.gz` file objects being created
for the same filesystem manager in consecutive iterations. The two ways we
avoid this, now, is by namespacing the data.tar.gz VFS but also by closing and
deleting after we're done.

An alternative would be to create a new FileSystemManager for each gem, but I'm
not sure if that's useful/worth it.